@@ -1,7 +1,5 @@
 package Utility;
 
-import java.awt.Point;
-
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.mygdx.game.Map.Map;
@@ -25,8 +23,7 @@ public class TiledMapPosition {
 	}
 
 	public TiledMapPosition setPositionFromScreen(float x, float y) {
-		tileCoordinates = new Point(Math.round(x / (Gdx.graphics.getWidth() / (float) BattleScreen.VISIBLE_WIDTH)),
-			Math.round(y / (Gdx.graphics.getHeight() / (float) BattleScreen.VISIBLE_HEIGHT)));
+		tileCoordinates = new Point(Math.round(x / (Gdx.graphics.getWidth() / (float) BattleScreen.VISIBLE_WIDTH)), Math.round(y / (Gdx.graphics.getHeight() / (float) BattleScreen.VISIBLE_HEIGHT)));
 		return this;
 	}
 
@@ -70,5 +67,16 @@ public class TiledMapPosition {
 	public TiledMapPosition changeY(int change) {
 		tileCoordinates.y += change;
 		return this;
+	}
+
+	private class Point {
+		private int x;
+		private int y;
+
+		public Point(int x, int y) {
+			this.x = x;
+			this.y = y;
+		}
+
 	}
 }
