@@ -28,7 +28,10 @@ import com.badlogic.gdx.graphics.g2d.freetype.FreeTypeFontGenerator;
 import com.badlogic.gdx.graphics.g2d.freetype.FreeTypeFontGeneratorLoader;
 import com.badlogic.gdx.graphics.g2d.freetype.FreetypeFontLoader;
 import com.badlogic.gdx.graphics.g2d.freetype.FreetypeFontLoader.FreeTypeFontLoaderParameter;
+import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.maps.tiled.TiledMap;
+import com.badlogic.gdx.math.Matrix4;
+import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.scenes.scene2d.ui.Label.LabelStyle;
 import com.mygdx.game.Entities.Entity;
 import com.mygdx.game.Entities.TeamLeader;
@@ -38,22 +41,17 @@ import com.mygdx.game.UI.MySkin;
 public final class Utility {
 	private static final String TAG = Utility.class.getSimpleName();
 
-	private static final String STATUSUI_TEXTURE_ATLAS_PATH = "skins/statusui.atlas";
-	private static final String UISKIN_TEXTURE_ATLAS_PATH = "skins/uiskin.atlas";
-	private static final String STATUSUI_SKIN_PATH = "skins/statusui.json";
-	private static final String SHADOW_WALKER_SKIN_TEXTURE_ATLAS_PATH = "skins/shadowWalker/shadow-walker-ui.atlas";
-	private static final String SHADOW_WALKER_SKIN_PATH = "skins/shadowWalker/shadow-walker-ui.json";
+	private static final String SKIN_TEXTURE_ATLAS_PATH = "skins/norii.atlas";
+	private static final String SKIN_JSON_PATH = "skins/norii.json";
 	public static final String ON_TILE_HOVER_FILE_PATH = "sprites/gui/selectedTile.png";
 
 	public static final AssetManager assetManager = new AssetManager();
 	private static InternalFileHandleResolver filePathResolver = new InternalFileHandleResolver();
+	private static ShapeRenderer debugRenderer = new ShapeRenderer();
 	public static final Random random = new Random();
 
 	private static TextureAtlas statusUITextureAtlas;
-	private static TextureAtlas uiTextureAtlas;
-	private static TextureAtlas shadowWalkerTextureAtlas;
 	private static MySkin statusUISkin;
-	private static MySkin shadowWalkerSkin;
 
 	public static void loadMapAsset(final String mapFilenamePath) {
 		loadAsset(mapFilenamePath, TiledMap.class, new MyNavTmxMapLoader(filePathResolver));
@@ -168,44 +166,20 @@ public final class Utility {
 		}
 	}
 
-	public static TextureAtlas getStatusUITextureAtlas() {
+	public static TextureAtlas getSkinTextureAtlas() {
 		if (statusUITextureAtlas == null) {
-			statusUITextureAtlas = new TextureAtlas(STATUSUI_TEXTURE_ATLAS_PATH);
+			statusUITextureAtlas = new TextureAtlas(SKIN_TEXTURE_ATLAS_PATH);
 		}
 
 		return statusUITextureAtlas;
 	}
 
-	public static TextureAtlas getUITextureAtlas() {
-		if (uiTextureAtlas == null) {
-			uiTextureAtlas = new TextureAtlas(UISKIN_TEXTURE_ATLAS_PATH);
-		}
-
-		return uiTextureAtlas;
-	}
-
-	public static TextureAtlas getShadowWalkerTextureAtlas() {
-		if (shadowWalkerTextureAtlas == null) {
-			shadowWalkerTextureAtlas = new TextureAtlas(SHADOW_WALKER_SKIN_TEXTURE_ATLAS_PATH);
-		}
-
-		return shadowWalkerTextureAtlas;
-	}
-
-	public static MySkin getStatusUISkin() {
+	public static MySkin getSkin() {
 		if (statusUISkin == null) {
-			statusUISkin = new MySkin(Gdx.files.internal(STATUSUI_SKIN_PATH), getStatusUITextureAtlas());
+			statusUISkin = new MySkin(Gdx.files.internal(SKIN_JSON_PATH), getSkinTextureAtlas());
 		}
 
 		return statusUISkin;
-	}
-
-	public static MySkin getShadowWalkersUISkin() {
-		if (shadowWalkerSkin == null) {
-			shadowWalkerSkin = new MySkin(Gdx.files.internal(SHADOW_WALKER_SKIN_PATH), getShadowWalkerTextureAtlas());
-		}
-
-		return shadowWalkerSkin;
 	}
 
 	public static float loadCompleted() {
@@ -322,5 +296,25 @@ public final class Utility {
 
 	private Utility() {
 
+	}
+
+	public static void DrawDebugLine(Vector2 start, Vector2 end, int lineWidth, Color color, Matrix4 projectionMatrix) {
+		Gdx.gl.glLineWidth(lineWidth);
+		debugRenderer.setProjectionMatrix(projectionMatrix);
+		debugRenderer.begin(ShapeRenderer.ShapeType.Line);
+		debugRenderer.setColor(color);
+		debugRenderer.line(start, end);
+		debugRenderer.end();
+		Gdx.gl.glLineWidth(1);
+	}
+
+	public static void DrawDebugLine(Vector2 start, Vector2 end, Matrix4 projectionMatrix) {
+		Gdx.gl.glLineWidth(2);
+		debugRenderer.setProjectionMatrix(projectionMatrix);
+		debugRenderer.begin(ShapeRenderer.ShapeType.Line);
+		debugRenderer.setColor(Color.DARK_GRAY);
+		debugRenderer.line(start, end);
+		debugRenderer.end();
+		Gdx.gl.glLineWidth(1);
 	}
 }
