@@ -6,10 +6,12 @@ public class MapFactory {
 	private static EnumMap<MapType, Map> mapTable = new EnumMap<MapType, Map>(MapType.class);
 
 	public enum MapType {
-		BATTLE_MAP
-	}
-
-	public enum MapFileNames {
+		BATTLE_MAP_THE_TOWN {
+			@Override
+			public String toString() {
+				return "maps/battle-raining.tmx";
+			}
+		},
 		BATTLE_MAP_THE_VILLAGE {
 			@Override
 			public String toString() {
@@ -20,17 +22,12 @@ public class MapFactory {
 
 	public static Map getMap(MapType mapType) {
 		Map map = null;
-		switch (mapType) {
-		case BATTLE_MAP:
-			map = mapTable.get(MapType.BATTLE_MAP);
-			if (map == null) {
-				map = new BattleMap(MapFileNames.BATTLE_MAP_THE_VILLAGE.toString());
-				mapTable.put(MapType.BATTLE_MAP, map);
-			}
-			break;
-		default:
-			break;
+		map = mapTable.get(mapType);
+		if (map == null) {
+			map = new BattleMap(mapType.toString());
+			mapTable.put(mapType, map);
 		}
+
 		return map;
 	}
 }
