@@ -1,12 +1,12 @@
 package com.mygdx.game.Particles;
 
 public enum ParticleType {
-	SPAWN("particles/mysmalleffect.p"),
-	MOVE("particles/mysmalleffect.p"),
-	SPELL("particles/mysmalleffect.p"),
-	FIREBALL("particles/fireball.p"),
-	SWAP("particles/swap.p"),
-	ATTACK("particles/attackParticle.p");
+	SPAWN("particles/spawn.p"), //
+	MOVE("particles/move.p"), //
+	SPELL("particles/mysmalleffect.p"), //
+	FIREBALL("particles/fireball.p"), //
+	SWAP("particles/swap.p"), //
+	ATTACK("particles/attack.p");//
 
 	private String particleFileLocation;
 
