@@ -8,15 +8,25 @@ public enum ScreenEnum {
 	MAIN_MENU {
 		@Override
 		public Screen getScreen(Object... params) {
-			return new MainMenuScreen();
+			final Screen main = ScreenManager.getMainMenu();
+			if (main == null) {
+				return new MainMenuScreen();
+			} else {
+				return ScreenManager.getMainMenu();
+			}
 		}
 	},
 	BATTLE {
 		@Override
 		public Screen getScreen(Object... params) {
-			@SuppressWarnings("unchecked")
 			final AITeams aiTeams = (AITeams) params[ScreenEnum.ScreenParams.AI_TEAM.ordinal()];
 			return new BattleScreen(aiTeams);
+		}
+	},
+	SETTINGS {
+		@Override
+		public Screen getScreen(Object... params) {
+			return new SettingsScreen();
 		}
 	};
 
