@@ -7,7 +7,6 @@ import com.badlogic.gdx.scenes.scene2d.ui.Cell;
 import com.badlogic.gdx.scenes.scene2d.ui.ImageButton;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
 import com.badlogic.gdx.scenes.scene2d.ui.Window;
-import com.mygdx.game.Entities.Entity;
 
 import Utility.Utility;
 
@@ -23,34 +22,26 @@ public abstract class UIWindow extends Window {
 	protected static final int ALPHA = 80;
 	protected static final float BUTTON_WIDTH_FACTOR = 20;
 	protected static final float BUTTON_HEIGHT_FACTOR = 20;
-	protected static final Color GOOD_COLOR = Color.GREEN;
-	protected static final Color BAD_COLOR = Color.RED;
 
 	protected UIWindow(String name, float width, float height) {
-		super(name, Utility.getStatusUISkin());
+		super(name, Utility.getSkin());
 
 		initVariables(width, height);
-		setFadeBackgroundEffect();
+		// setFadeBackgroundEffect();
 	}
 
 	protected void initVariables(float width, float height) {
 		tileWidthPixel = Gdx.graphics.getWidth() / (float) TILE_TO_PIXEL_RATIO;
 		tileHeightPixel = Gdx.graphics.getHeight() / (float) TILE_TO_PIXEL_RATIO;
-		this.windowWidth = width;
-		this.windowHeight = height;
+		windowWidth = width;
+		windowHeight = height;
+		// final WindowStyle styleTransparent = Utility.getSkin().get("pixthulhu", WindowStyle.class);
+		// setStyle(styleTransparent);
 	}
 
 	private void setFadeBackgroundEffect() {
 		final Color newColor = this.getColor();
 		newColor.a = ALPHA;
-	}
-
-	protected void setBackgroundColor(Window window, Entity entity) {
-		if (entity.isPlayerUnit()) {
-			window.setColor(GOOD_COLOR);
-		} else {
-			window.setColor(BAD_COLOR);
-		}
 	}
 
 	public void update() {
